requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asyncpg==0.30.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
databases==0.9.0
DateTime==5.5
distlib==0.3.9
Django==5.0
django-appconf==1.1.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-geoip==0.5.2.1
django-sms==0.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
fastapi==0.115.11
fastapi-jwt-auth==0.5.0
filelock==3.17.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.10
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
progressbar2==4.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.1
requests==2.32.3
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.2
VirtualEnvironment==0.1.2
zope.interface==7.2
